refactor(ui): extract settings/data/collect-payload.ts — pure form→payload reducer
